requirements.txt
click==7.1.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mongoengine==0.19.1
passlib==1.7.2
PyJWT==1.7.1
pymongo==3.10.1
python-dotenv==0.12.0
six==1.14.0
Werkzeug==1.0.1
